semoga awet

+++ app/Libs/Helpers.php
@@ -4,9 +4,25 @@ namespace App\Libs;
 use Image;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\DB;
 
 class Helpers
 {
+  public static function prefix($data)
+  {
+    
+    $prefix = DB::table('_regencies')
+      ->where('province_id', '15')
+      ->where('id', $data->regencies_id)
+      ->select(["id", DB::Raw("
+      Case WHEN id = '1501' THEN 'KRC'
+        WHEN id = '1572' THEN 'SPN'
+        ELSE 'NaN' END as prefix
+      ")])
+      ->first();
+    return $prefix;
+  }
+
   public static function prepareFile($inputs, $subFolder)
   {
     $file = new \StdClass;
@@ -39,7 +55,30 @@ class Helpers
     return $file;
   }
 
-  public static function prepareFile1($inputs, $subFolder)
+  public static function prepareAnggota($inputs, $subFolder)
+  {
+    $file = new \StdClass;
+    try {
+      $file = isset($inputs['file']) ? $inputs['file'] : null;
+      $path = 'public' . $subFolder;
+        if (!File::isDirectory($path)) {
+          Storage::makeDirectory($path);
+        }
+      if($file != null){
+        $file->path = storage_path('app/public') . $subFolder;
+        $file->newName = time()."_".$inputs['nama_lengkap'].'.'.$file->getClientOriginalExtension();
+        $file->originalName = explode('.',$file->getClientOriginalName())[0];
+        // $file->move($file->path ,$file->newName);
+        Image::make($file)->resize(332, 500)->save($file->path. "/" . $file->newName);
+      }
+    } catch (\Exception $e){
+      dd($e);
+        // supress
+    }
+    return $file;
+  }
+
+  public static function prepareKtp($inputs, $subFolder)
   {
     $file = new \StdClass;
     try {
@@ -50,19 +89,11 @@ class Helpers
         }
       if($file != null){
         $file->path = storage_path('app/public') . $subFolder;
-        $file->newName = time()."_".$file->getClientOriginalName();
+        $file->newName = time()."_".$inputs['nama_lengkap'].'.'.$file->getClientOriginalExtension();
         $file->originalName = explode('.',$file->getClientOriginalName())[0];
         // $file->move($file->path ,$file->newName);
-        Image::make($file)->save($file->path. "/" . $file->newName);
+        Image::make($file)->resize(300, 199)->save($file->path. "/" . $file->newName);
 
-        //buat folder tumbnail
-        $critPath = $path .'/'. 'thumbnail/';
-        if (!File::isDirectory($critPath)) {
-          Storage::makeDirectory($critPath);
-        }
-        $tumbPath = $file->path .'/'. 'thumbnail/';
-        $img = Image::make($file);
-        $img->resize(350, 220)->save($tumbPath . "/" . $file->newName);
       }
     } catch (\Exception $e){
       dd($e);
@@ -75,7 +106,8 @@ class Helpers
   {
     if(Storage::exists('public/'.$subFolder.'/'.$inputs)){
       Storage::delete('public/'.$subFolder.'/'.$inputs);
-      Storage::delete('public/'.$subFolder.'/'.'thumbnail/'.$inputs);
+      if(Storage::exists('public/'.$subFolder.'/'.'thumbnail/'.$inputs))
+        Storage::delete('public/'.$subFolder.'/'.'thumbnail/'.$inputs);
     }
   }
